scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.28.0",
    "pydantic>=2.11.7",
]
